Added Command Class
